pyproject.toml
[project]
name = "cognitex"
version = "0.1.0"
description = "Personal agent system for cognitive load management"
readme = {text = "Personal agent system for cognitive load management", content-type = "text/plain"}
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Databases
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "neo4j>=5.26.0",
    "redis>=5.2.0",

    # Background tasks
    "arq>=0.26.1",

    # LLM
    "together>=1.3.0",
    "httpx>=0.28.0",

    # Google APIs
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.154.0",

    # Discord
    "discord.py>=2.4.0",

    # CLI
    "textual>=0.89.0",
    "rich>=13.9.0",
    "typer>=0.14.0",

    # Utilities
    "python-dateutil>=2.9.0",
    "structlog>=24.4.0",

    # Document extraction
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
cognitex = "cognitex.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cognitex"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["cognitex"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra -q"
